Submit on enter key
Functionality added + style&bug fixes

+++ frontend/src/app/modals/delete-account/delete-account.component.ts
@@ -1,4 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
+import { NgForm } from '@angular/forms';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ToastrService } from 'ngx-toastr';
 import { ApiService } from 'src/app/Services/api.service';
@@ -32,15 +33,36 @@ export class DeleteAccountComponent {
   /**
  * Deletes The Seed-Test account
  */
-  deleteAccount() {
-    const insertedEmail = (document.getElementById('insertedEmail') as HTMLInputElement).value;
-    if (insertedEmail == this.email) {
+  deleteAccount(form: NgForm) {
+    const insertedEmail = form.value.insertedEmail;
+    
+    if (insertedEmail !==undefined && insertedEmail === this.email) {
         this.apiService.deleteUser().subscribe(resp => {
             this.toastr.info('', 'User Deleted');
             this.apiService.logoutEvent.emit();
         });
     } else {
-        this.modalService.open(this.deleteAccountModal, {ariaLabelledBy: 'modal-basic-title', size: 'sm'});
+      this.unallowableNameToast();
+      this.modalService.open(this.deleteAccountModal, {ariaLabelledBy: 'modal-basic-title', size: 'sm'});
     }
+  }
+
+  /**
+ * Opens warning toast
+ */
+
+  unallowableNameToast() {
+    this.toastr.warning('', 'Enter a valid e-mail', {
+    });
+  }
+
+  enterSubmit(event, form: NgForm) {
+    if (event.keyCode === 13) {
+      this.deleteAccount(form);
+    }
+  }
+
+  onClickSubmit(form: NgForm) {
+    this.deleteAccount(form);
   }
 }
